[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17692 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         
-        int result = (a + b) * (a - b);
+        long result = (a + b) * (a - b);
         
         System.out.println(result);
     }
